+++ labbook_BE/default_settings.py
@@ -1,6 +1,6 @@
 # -*- coding:utf-8 -*-
 
-APP_VERSION = '3.6.2'
+APP_VERSION = '3.6.3'
 
 # ######################################
 #  Database parameters